chore(SCO-2): get tests to green
