+++ internal/config/tokens.go
@@ -1,0 +1,6 @@
+package config
+
+const (
+	SPACE = " "
+	COLON = ":"
+)
