+++ internal/invoke/httpoke/patch.go
@@ -1,9 +1,10 @@
-package invoke
+package httpoke
 
 import (
 	"bytes"
 	"context"
 	"fmt"
+	"github.com/dark-enstein/scour/internal/invoke"
 	"github.com/dark-enstein/scour/internal/parser"
 	"github.com/dark-enstein/scour/internal/parser/httparser"
 	"io"
@@ -12,14 +13,13 @@ import (
 	"time"
 )
 
-// Get sends a GET HTTP request to the specified URL.
-// It manages request timeouts using context, logs relevant information,
-// and returns the response headers and body as a byte slice.
-func Get(ctx context.Context, url parser.Url) (*RespHeaders, []byte, error) {
-	_ = &RespHeaders{}
+// Patch sends a PATCH HTTP request to the specified URL with the provided data.
+// It returns the response headers and body.
+func Patch(ctx context.Context, url parser.Url, data []byte) (*invoke.RespHeaders, []byte, error) {
+	_ = &invoke.RespHeaders{}
 	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
 	defer cancel()
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url.String(), bytes.NewBuffer([]byte{}))
+	req, err := http.NewRequestWithContext(ctx, http.MethodPatch, url.String(), bytes.NewBuffer(data))
 	if err != nil {
 		log.Printf("Error creating request object: %s\n", err.Error())
 	}
@@ -28,7 +28,7 @@ func Get(ctx context.Context, url parser.Url) (*RespHeaders, []byte, error) {
 	t1 := time.Now()
 	resp, err := cli.Do(req)
 	if err != nil {
-		log.Printf("GET request failed with: %s\n", err.Error())
+		log.Printf("POST request failed with: %s\n", err.Error())
 		return nil, nil, err
 	}
 	defer func(Body io.ReadCloser) {
@@ -36,7 +36,7 @@ func Get(ctx context.Context, url parser.Url) (*RespHeaders, []byte, error) {
 	}(resp.Body)
 	tDur := time.Since(t1)
 
-	respH := newHeaders(resp.Status, fmt.Sprintf("%s/1.1", url.Protocol().String()), resp.Header.Get("Date"), resp.Header.Get("Content-Type"), resp.Header.Get("Content-Length"), resp.Header.Get("Connection"), resp.Header.Get("Server"), resp.Header.Get("Access-Control-Allow-Origin"), resp.Header.Get("Access-Control-Allow-Credentials"))
+	respH := invoke.NewHeaders(resp.Status, fmt.Sprintf("%s/1.1", url.Protocol().String()), resp.Header.Get("Date"), resp.Header.Get("Content-Type"), resp.Header.Get("Content-Length"), resp.Header.Get("Connection"), resp.Header.Get("Server"), resp.Header.Get("Access-Control-Allow-Origin"), resp.Header.Get("Access-Control-Allow-Credentials"))
 	if httparser.ParseLogLevelFromCtx(ctx, httparser.KeyV) == true {
 		log.Printf("Response: %v\n", respH)
 	}
@@ -57,6 +57,5 @@ func Get(ctx context.Context, url parser.Url) (*RespHeaders, []byte, error) {
 	if httparser.ParseLogLevelFromCtx(ctx, httparser.KeyV) == true {
 		log.Printf("Time taken: %s\n", tDur.String())
 	}
-
 	return respH, responseStream, nil
 }
